// client/src/api/generated/models/VehicleResponseDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type VehicleResponseDto = {
    /**
     * Display name for the vehicle
     */
    name: string;
    /**
     * Unique licence plate number
     */
    licensePlate: string;
    id: string;
    createdAt: string;
};